// apps/frontend/src/app/core/api/models/column-config-input.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { CastType } from '../models/cast-type';
import { ColumnFilter } from '../models/column-filter';

/**
 * Column configuration for transformation.
 */
export interface ColumnConfigInput {

  /**
   * Target data type for casting. None = keep original type.
   */
  cast_type?: (CastType | null);

  /**
   * Whether the column is excluded from output
   */
  excluded?: boolean;

  /**
   * Active filter on this column. None = no filter.
   */
  filter?: (ColumnFilter | null);

  /**
   * Renamed column name. None = keep original name.
   */
  new_name?: (string | null);

  /**
   * Original column name from staging table (immutable reference)
   */
  original_name: string;
}