// apps/frontend/src/app/core/api/models/staging-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { DagRunState } from '../models/dag-run-state'

/**
 * Response model for import endpoint
 */
export interface StagingResponse {
  dag_id: string
  dag_run_id: string
  integrity_link_id: string
  status: DagRunState
}