// apps/frontend/src/app/core/api/models/staging-preview-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { FeatureCollectionFeatureAnnotatedUnionPointMultiPointLineStringMultiLineStringPolygonMultiPolygonGeometryCollectionFieldInfoAnnotationNoneTypeRequiredTrueDiscriminatorTypeDictStrAny } from '../models/feature-collection-feature-annotated-union-point-multi-point-line-string-multi-line-string-polygon-multi-polygon-geometry-collection-field-info-annotation-none-type-required-true-discriminator-type-dict-str-any'

/**
 * Preview data from staging table
 */
export interface StagingPreviewResponse {
  data: Array<{
    [key: string]: any
  }>

  /**
   * GeoJSON FeatureCollection, only present for geographic data
   */
  geojson?: FeatureCollectionFeatureAnnotatedUnionPointMultiPointLineStringMultiLineStringPolygonMultiPolygonGeometryCollectionFieldInfoAnnotationNoneTypeRequiredTrueDiscriminatorTypeDictStrAny | null

  /**
   * Indicates if the staging table contains geometry data
   */
  is_geographic?: boolean
}